Finish README
Added Section:
This is a real-time machine learning project designed with the intent of recognizing and identifying people and checking if the person is masked or not in real time. The program was run on a NVIDIA Jetson Nano which enabled the use of a real-time camera feed and display screen. It uses a Convolutional Neural Network (CNN) to differentiate between a masked and unmasked individual. There are other sections in the program that store images from a file as people who are known. If a person whose image is saved in the file appears before the camera, they are recognized and their name is displayed on the screen. If the person is not in the file with known people, they are labeled as unknown along with the number of times they have been detected on camera. The unknown person's image is saved into a separate folder with the number of times they have visited. This image is updated every time the person comes in front of the camera. With masked individuals, regardless of whether they are known or not, their images are saved into a file with a short list of names of people it could be.